#18 refactor: moved inner contents to organisms
